- Remove needless dereferences
- Add inline attributes


Benchmark results of `main` branch:
```
test movegen::bench_legal_moves_from_default ... bench:         934 ns/iter (+/- 212)
test movegen::bench_legal_moves_maximum      ... bench:       3,452 ns/iter (+/- 203)

test perft::bench_perft_3_from_maximum_moves ... bench: 276,255,127 ns/iter (+/- 11,050,061)
test perft::bench_perft_5_from_default       ... bench: 320,198,552 ns/iter (+/- 5,421,506)
```

Benchmark results of this branch:
```
test movegen::bench_legal_moves_from_default ... bench:         878 ns/iter (+/- 35)
test movegen::bench_legal_moves_maximum      ... bench:       3,347 ns/iter (+/- 84)

test perft::bench_perft_3_from_maximum_moves ... bench: 251,215,303 ns/iter (+/- 4,981,781)
test perft::bench_perft_5_from_default       ... bench: 280,564,996 ns/iter (+/- 3,973,011)
```